Split pyodide.js into several files
