+++ package.json
@@ -1,0 +1,6 @@
+{
+  "dependencies": {
+    "prettier": "^2.2.1",
+    "rollup-plugin-terser": "^7.0.2"
+  }
+}
