+++ src/templates/console.html
@@ -1,24 +1,30 @@
-<!doctype html>
+<!DOCTYPE html>
 <html>
   <head>
-    <meta charset="UTF-8">
+    <meta charset="UTF-8" />
     <script src="https://cdn.jsdelivr.net/npm/jquery"></script>
     <script src="https://cdn.jsdelivr.net/npm/jquery.terminal@2.23.0/js/jquery.terminal.min.js"></script>
-    <link href="https://cdn.jsdelivr.net/npm/jquery.terminal@2.23.0/css/jquery.terminal.min.css" rel="stylesheet"/>
+    <link
+      href="https://cdn.jsdelivr.net/npm/jquery.terminal@2.23.0/css/jquery.terminal.min.css"
+      rel="stylesheet"
+    />
     <script src="{{ PYODIDE_BASE_URL }}pyodide.js"></script>
     <style>
-      .terminal { --size: 1.5; }
+      .terminal {
+        --size: 1.5;
+      }
     </style>
   </head>
   <body>
     <script>
-      function sleep(s){
-        return new Promise(resolve => setTimeout(resolve, s));
+      function sleep(s) {
+        return new Promise((resolve) => setTimeout(resolve, s));
       }
       async function main() {
-        await loadPyodide({ indexURL : '{{ PYODIDE_BASE_URL }}' });
+        await loadPyodide({ indexURL: "{{ PYODIDE_BASE_URL }}" });
         let namespace = pyodide.globals.get("dict")();
-        pyodide.runPython(`
+        pyodide.runPython(
+          `
             import sys
             import js
             from pyodide import console
@@ -36,15 +42,18 @@
 
 
             js.pyconsole = PyConsole()
-        `, namespace);
+        `,
+          namespace
+        );
         namespace.destroy();
 
-        let ps1 = '>>> ', ps2 = '... ';
+        let ps1 = ">>> ",
+          ps2 = "... ";
 
-        async function lock(){
+        async function lock() {
           let resolve;
           let ready = term.ready;
-          term.ready = new Promise(res => resolve = res);
+          term.ready = new Promise((res) => (resolve = res));
           await ready;
           return resolve;
         }
@@ -54,22 +63,22 @@
           try {
             term.pause();
             // multiline should be splitted (useful when pasting)
-            for( const c of command.split('\n') ) {
+            for (const c of command.split("\n")) {
               let run_complete = pyconsole.run_complete;
               try {
-                  const incomplete = pyconsole.push(c);
-                  term.set_prompt(incomplete ? ps2 : ps1);
-                  let r = await run_complete;
-                  if(pyodide.isPyProxy(r)){
-                    r.destroy();
-                  }
-                } catch(e){
-                  if(e.name !== "PythonError"){
-                    term.error(e);
-                    throw e;
-                  }
+                const incomplete = pyconsole.push(c);
+                term.set_prompt(incomplete ? ps2 : ps1);
+                let r = await run_complete;
+                if (pyodide.isPyProxy(r)) {
+                  r.destroy();
                 }
-                run_complete.destroy();
+              } catch (e) {
+                if (e.name !== "PythonError") {
+                  term.error(e);
+                  throw e;
+                }
+              }
+              run_complete.destroy();
             }
           } finally {
             term.resume();
@@ -78,25 +87,24 @@
           }
         }
 
-        let term = $('body').terminal(
-          interpreter,
-          {
-            greetings: pyconsole.banner(),
-            prompt: ps1,
-            completionEscape: false,
-            completion: function(command, callback) {
-              callback(pyconsole.complete(command).toJs()[0]);
-            }
-          }
-        );
+        let term = $("body").terminal(interpreter, {
+          greetings: pyconsole.banner(),
+          prompt: ps1,
+          completionEscape: false,
+          completion: function (command, callback) {
+            callback(pyconsole.complete(command).toJs()[0]);
+          },
+        });
         window.term = term;
-        pyconsole.stdout_callback = s => term.echo(s, {newline : false});
-        pyconsole.stderr_callback = s => {
-            term.error(s.trimEnd());
-        }
+        pyconsole.stdout_callback = (s) => term.echo(s, { newline: false });
+        pyconsole.stderr_callback = (s) => {
+          term.error(s.trimEnd());
+        };
         term.ready = Promise.resolve();
         pyodide._module.on_fatal = async (e) => {
-          term.error("Pyodide has suffered a fatal error. Please report this to the Pyodide maintainers.");
+          term.error(
+            "Pyodide has suffered a fatal error. Please report this to the Pyodide maintainers."
+          );
           term.error("The cause of the fatal error was:");
           term.error(e);
           term.error("Look in the browser console for more details.");
