+++ src/templates/test.html
@@ -1,24 +1,24 @@
 <!-- Bootstrap HTML for running the unit tests. -->
 <html>
   <head>
-  <meta charset="utf-8"/>
-  <script type="text/javascript">
-    window.logs = [];
-    console.log = function(message) {
-      window.logs.push(message);
-    };
-    console.warn = function(message) {
-      window.logs.push(message);
-    };
-    console.info = function(message) {
-      window.logs.push(message);
-    };
-    console.error = function(message) {
-      window.logs.push(message);
-    };
-    // set the pyodide files URL (packages.json, pyodide.asm.data etc)
-  </script>
-  <script src="./pyodide.js"></script>
-</head>
-<body></body>
+    <meta charset="utf-8" />
+    <script type="text/javascript">
+      window.logs = [];
+      console.log = function (message) {
+        window.logs.push(message);
+      };
+      console.warn = function (message) {
+        window.logs.push(message);
+      };
+      console.info = function (message) {
+        window.logs.push(message);
+      };
+      console.error = function (message) {
+        window.logs.push(message);
+      };
+      // set the pyodide files URL (packages.json, pyodide.asm.data etc)
+    </script>
+    <script src="./pyodide.js"></script>
+  </head>
+  <body></body>
 </html>
